Update for Serde v1.0
This patch also fixes a longstanding bug in the way we deserialize
author and category listings. The bug never mattered before because the
API doesn't *really* support getting those (and anyway they're only
available to premium members), but basically the API response for those
calls returns a string instead of a number as the "total" available.

...Anyway, instead of using a single `ApiResponse` wrapper, we have a
standard response wrapper and then one response struct each for Author
and for Category. Which is stupid as hell, but I don't really care to
tell them how to run their business, I guess.

Hell, it's possible they've changed this API since the last time I
checked; it's not as if I'm going to give them another months'
membership fees to find out.

.......Anyway, I don't really even know how that was being handled
before--I don't know if the tests that covered that just never worked
before (the build is listed as passing?!) or if maybe the older version
of serde just went ahead and parsed a string as an integer if you asked
it to or what, but for *those* API endpoints, we now parse the "number"
as an `Option(i32)` because I have no idea what to show our users if
they send back `"lots"` instead of `1513297`.